ivshmem: add support for exposing pmd ring ethernet devices to the guest
the host application can expose to the guest a virtual ethernet device
by means of the rte_ivshmem_metadata_add_pmd_ring function, the host
provides the rings that are used be the guest as queues of the device.

When the guest application starts, the EAL automatically detects the virtual
device and adds it to the list of devices, then the application can use it
by means of the rte_ethdev library.