add special pmd_ring driver
This commit modifies the ring driver in order to allow it to
use a slave device. When the slave device is added, the driver
always sends packets using this device but it still continues to
read both, slave and primary devices.

When the slave device is removed, the driver returns back to its
normal behaviour.

In the case of direct vm2vm, add slave device will be created and
attached to the vm when a direct path is detected, the fact that there
is still a part of the device that is connected to the datapath allows
the virtual swith to send packets to the application.