ivshmem: create function to get pmdinternals structure
when a pmd_ring is added to an ivshmme device, it creates a internal pmd_structure
that contains the rings, the statistics and other information for the device.
It is important for an application to have direct access with this in order to control
the pmd ring running in the guest.