virtio_serial: use dedicated thread for reading device
Previous implementation set the virtio serial device to async mode, then a
signal handler was used to read from the device the data arrrived.
This implementation presented some issues because the function to
configure the new device was called inside the main thread, it caused
the packet processing of the app to be delayed.

This commit adds a new thread that reads the virtio serial and does the
bypass device configuration.