pmd_ring: send single cap packet
A previous strategy sent multiple cap packets to avoid losing them.
This strategy actually generates more problems than solutions, if a packet
is lost it is very probable that packets that are near are also lost.
Another issue is that a special mechanisms to avoid the cap packets
reaching the application is needed when sending multiple...